Refactors Lean Report Creator
`LeanReportCreator` is now the HTML creator and `LeanOutputReport` reads a dictionary (the Lean output) to create the images and statistics.
