+++ PythonToolbox/QuantConnect.PythonToolbox.pyproj
@@ -22,8 +22,9 @@
     <EnableUnmanagedDebugging>false</EnableUnmanagedDebugging>
   </PropertyGroup>
   <ItemGroup>
+    <Compile Include="CreateLeanReport.py" />
     <Compile Include="quantconnect\api.py" />
-    <Compile Include="quantconnect\GenerateHTML.py" />
+    <Compile Include="quantconnect\LeanOutputReader.py" />
     <Compile Include="quantconnect\LeanReportCreator.py" />
     <Compile Include="quantconnect\__init__.py" />
     <Compile Include="setup.py" />
@@ -37,10 +38,12 @@
     <Folder Include="quantconnect\" />
   </ItemGroup>
   <ItemGroup>
+    <Content Include="AuthorProfile.png" />
     <Content Include="MANIFEST.in" />
     <Content Include="README.rst">
       <SubType>Code</SubType>
     </Content>
+    <Content Include="user_data.json" />
   </ItemGroup>
   <Import Project="$(MSBuildExtensionsPath32)\Microsoft\VisualStudio\v$(VisualStudioVersion)\Python Tools\Microsoft.PythonTools.targets" />
   <!-- Uncomment the CoreCompile target to enable the Build command in
