--- outputs/strategy-statistics.json
@@ -1,1 +1,0 @@
-{"Key Characteristics": [["Significant Period", 1], ["Significant Trading", 1], ["Diversified", 0], ["Risk Control", 1], ["Markets", ["Equity"]]], "Key Statistics": [["CAGR", "24.06%"], ["Drawdown", "12.7%"], ["Sharpe Ratio", 1.783], ["Information Ratio", 0.701], ["Trades Per Day", 2.367396]]}
